pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=80.0"]

[project]
name = "batch_img"
version = "0.0.7"
description = "Batch processing image files by utilizing Pillow / PIL library"
readme = "README.md"
authors = [{ name = "John Liu" }]
license-files = ["LICENSE"]
classifiers = [
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: 3 :: Only",
    "Operating System :: OS Independent",
]
requires-python = ">=3.12"
dependencies = [
    "pillow",
    "click",
    "loguru",
]
[project.scripts]
batch_img = "batch_img.interface:cli"

[project.urls]
Download = "https://artifacts.some.where"
Homepage = "https://github.com/john-liu2/batch_img"

[tool.setuptools.packages.find]
include = ["*"]

# Dev config
[project.optional-dependencies]
# uv pip compile pyproject.toml -o requirements.txt
# uv pip install -e '.[dev]'
dev = [
    "pytest-cov", # also install pytest
    "black>=25.1.0",
    "ruff",
    "pre-commit", # rules are in .pre-commit-config.yaml
    "pylint",
]

[tool.ruff]
line-length = 88
lint.select = [
    "B", # bugbear
    "E", # pycodestyle errors
    "F", # pyflakes
    "I", # isort
    "UP", # pyupgrade
    "PD", # pandas-vet
    "RET", # flake8-return (includes no-else-return)
    "PL", # pylint
    "C416", # Unnecessary comprehension (R1721 from Pylint)
]
lint.ignore = [
    "RET504",
    "PLR0913",
    "PLR0917",
    "PLR0904",
    "PLR1702", # Too many nested blocks
    #    "PLR0801",
    "PLR2004", # Magic value used in comparison, consider replacing `3` with a constant variable
    "PLC0206", # Extracting value from dictionary without calling `.items()`
    "PLR5501", # Use `elif` instead of `else` then `if`, to reduce indentation
    "PLR6301", # Method `is_xxx` could be a function, class method, or static method
    "E266", # Too many leading `#` before block comment
]
lint.fixable = ["E", "F", "I", "UP"]
preview = true
exclude = [
    "**/venv/",
    "**/__pycache__/",
    "**/migrations/",
    "**/tests/",
    "**/build/",
    "**/generated/**/*.py",
]
